requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
metatrader5==5.0.5735
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2026.1
streamlit
plotly
fastapi
uvicorn
bcrypt
python-jose[cryptography]
mysql-connector-python
python-multipart
